Online Student Scheduling: XE Interface
- get assigned (section) credit from the Banner's enrollment data and display it in the Scheduling Assistant

+++ JavaSource/org/unitime/timetable/onlinesectioning/custom/purdue/XEInterface.java
@@ -69,7 +69,7 @@ public class XEInterface {
 		 */
 		public String courseRegistrationStatus;
 		public String courseRegistrationStatusDescription;
-		public Double creditHour;
+		public Float creditHour;
 		
 		public String gradingMode;
 		public String gradingModeDescription;
